something

+++ WindowsProject1/Vector2.cpp
@@ -4,7 +4,7 @@ Vector2::Vector2() {
 	x = y = 0;
 }
 
-Vector2::Vector2(float x, float y) {
+Vector2::Vector2(double x, double y) {
 	this->x = x;
 	this->y = y;
 }
