+++ WindowsProject1/Vector2.h
@@ -4,11 +4,11 @@
 class Vector2
 {
 public:
-	float x;
-	float y;
+	double x;
+	double y;
 
 	Vector2();
-	Vector2(float x, float y);
+	Vector2(double x, double y);
 
 	Vector2* Normalized();
